+++ tests/testthat/test_RSL.R
@@ -1,10 +1,13 @@
+set.seed(123)
 library(Bchron)
+library(vdiffr)
 library(ggplot2)
-set.seed(123) # Added in to avoid weird bug on GitHub Actions
+co <- function(expr) capture.output(expr, file = "NUL")
+
 data(TestChronData)
 data(TestRSLData)
 
-RSLchron <- with(
+co(RSLchron <- with(
   TestChronData,
   Bchronology(
     ages = ages,
@@ -19,8 +22,8 @@ RSLchron <- with(
     burn = 20,
     thin = 1
   )
-)
-RSLrun <- with(
+))
+co(RSLrun <- with(
   TestRSLData,
   BchronRSL(RSLchron,
     RSLmean = RSL,
@@ -30,7 +33,7 @@ RSLrun <- with(
     burn = 20,
     thin = 1
   )
-)
+))
 
 test_that("RSL Bchronology", {
   expect_s3_class(RSLchron, "BchronologyRun")
@@ -41,14 +44,14 @@ test_that("RSL functions", {
 })
 
 test_that("summary and plot RSL functions", {
-  expect_s3_class(
-    plot(RSLrun, type = "RSL") + ggtitle("Relative sea level plot"),
-    "ggplot"
-  )
-  expect_s3_class(plot(RSLrun, type = "rate") + ggtitle("Rate of RSL change") +
-    ylab("Rate (mm per year)"), "ggplot")
-  expect_s3_class(plot(RSLrun, type = "accel") + ggtitle("Rate of RSL change") +
-    ylab("Rate (mm per year)"), "ggplot")
+  p <- plot(RSLrun, type = "RSL") + ggtitle("Relative sea level plot")
+  expect_doppelganger("RSL_p1", p)
+  p <- plot(RSLrun, type = "rate") + ggtitle("Rate of RSL change") +
+    ylab("Rate (mm per year)")
+  expect_doppelganger("RSL_p2", p)
+  p <- plot(RSLrun, type = "accel") + ggtitle("Rate of RSL change") +
+    ylab("Rate (mm per year)")
+  expect_doppelganger("RSL_p3", p)
   expect_output(summary(RSLrun, type = "RSL", age_grid = seq(0, 2000, by = 250)))
   expect_output(summary(RSLrun, type = "parameters", age_grid = seq(0, 2000, by = 250)))
   expect_output(summary(RSLrun, type = "rate", age_grid = seq(0, 2000, by = 250)))